Remastered

+++ src/com/company/GameResult.java
@@ -3,5 +3,5 @@ package com.company;
 public class GameResult {
     String userName;
     int attempts;
-    double time;
+    long time;
 }
